Detecting aruco markers
