+++ camera/requirements.txt
@@ -1,1 +1,2 @@
 opencv-python
+requests
